Object 3 made please feel free to change up as this is just a test to change the flow.
Made 3rd object mimic Object2 , but changed the gravity with a binary Digit from text which adds to Object3's xx_vel and yy_vel in the Timer section and I reversed the  it where the first if is + the second if is - the 3rd if is - and the fourth if is + with respect to adding Convert which is from the Binary string. 

I reversed the rotational pull as in make it go clockwise then changes the rotation to counter clockwise for Object 2 and Object 3.